v1.4 Admin Page

+++ lib/utils.ts
@@ -17,7 +17,7 @@ export function getInitials(name: string): string {
 export function formatDate(date: Date): string {
   return new Date(date).toLocaleDateString("en-US", {
     year: "numeric",
-    month: "long",
+    month: "short",
     day: "numeric",
   });
 };
